Add div .container to every section of page
I added a div with a class of .container to every section of webpage in
HTML "index.html" file to better handel the CSS styles followed by the
related changes in CSS "styles.css" file for example i set the
"div . container" width to 1100 px.
I also added an SVG file as an image in hero section of the "index.html" page.
I also modified the CSS file a bit to have a better looking and more
professional code.